ARM: ux500: Allocate correct amount of memory for the u8540 in DT
As it stands, the memory node in the u8540 Device Tree is ignored and
memory is actually stipulated by the bootloader via ATAGS. ARM core
architecture code then extracts the memory ATAG and inserts it into the
Device Tree. In the u8540 the LittleKernel bootloader only stipulates
512MB of memory; however, the u8540 actually has 2GB. We're taking the
responsibility to register the remainder from Device Tree here.

Signed-off-by: Lee Jones <lee.jones@linaro.org>
Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 2388af4e9b190cef9fb24e49654010fcbd0ecf97 ]