Merge remote-tracking branch 'agust/next' into next
Fixes for build issues when LPB FIFO driver is configured as
a module, removal of #ifdefs in mpc512x DIU platform code and
a revert of recent changes to mpc52xx PIC driver. Wolfram
provided a better fix for PIC driver build issue popping up
when older gcc-4.3.5 is used.

[ upstream commit: 5c509a2bb4d3ea6a26393e13ac4f054a2fb111ca ]