ARM: ux500: move the WLAN GPIO pin setup to the device tree
This moves some of the pin setup related to the CW1200 WLAN
module over to the device tree. As the driver is not yet
activated for the CW1200 WLAN we do not assign this pinctrl
state to any device node yet.

Get rid of the cmdline argument passing of a certain U9500
platform variant, as this is not supported by the kernel or
any device tree.

Signed-off-by: Linus Walleij <linus.walleij@linaro.org>

[ upstream commit: 5026119fbef49ce64fc5469c5d3c2d7c313469fb ]