Merge branch 'macvlan'
Michal Kubecek says:

====================
macvlan: disable LRO on lowerdev instead of a macvlan

A customer of ours encountered a problem with LRO on an ixgbe network
card. Analysis showed that it was a known conflict of forwarding and LRO
but the forwarding was enabled in an LXC container where only a macvlan
was, not the ethernet device itself.

I believe the solution is exactly the same as what we do for "normal"
(802.1q) VLAN devices: if dev_disable_lro() is called for such device,
LRO is disabled on the underlying "real" device instead.

v2: adapt to changes merged from net-next

v3: use BUG() in macvlan_dev_real_dev() if compiled without macvlan
====================

Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 82c80e9d9de354333aa69368576df6a355791418 ]